Automatic ability importing.
Importing a rank from the wiki now automatically imports abilities for that rank into the compendium the rank is located in. The rank also maintains a list of links to the abilities that can be referenced on its sheet. These function even when the rank is on an actor or in the world and the ability is in the compendium.