Style improvements and mechanics removed.
Automatic encumbrance and wounded states are no longer handled by a separate item with active effects. Instead, they are managed synchronously as "virtual effects". These are not real active effects, but just data that appears as an active effect on tokens and in actor sheets. This modifies actor post updates and token data quite substantially and might introduce bugs in those areas.

Style improvements mainly revolve around unifying cards with images in other places. Card images now have a similar border to sheet headers and panels. Instead of a "marker" being placed to the left of cards, colors are implemented as a border on the images themselves.